Update features, Printing
# Update features
Now deleting node with two children.
Update deleting a node with 1 value. now exchange value, like the deletion of a node with 2 children. It has been modified because the parent of a child never change, so by logic, it has been change to resemble the deletion of a node with 2 children.
Update is_leaf to avoid a bug, that occurred when deleting a node after the exchange of value of node with 1 and 2 children

# Printing
Add new tests

## IMPORTANT
Balancing has been temporarily abandoned to carry out the basic project.It may be added later.

So this file is finished. Next, transform a tree into a tree list